requirements.txt
discord.py
PyNaCl
typing_extensions
pafy
opuslib
BeautifulSoup
urllib
pandas
requests